Tweakcraft changes for Minecraft 1.8. UUID Support not yet implemented (fully).

+++ src/com/zones/accessresolver/interfaces/PlayerAttackEntityResolver.java
@@ -6,5 +6,5 @@ import org.bukkit.entity.Player;
 import com.zones.model.ZoneBase;
 
 public interface PlayerAttackEntityResolver extends MessagebleResolver {
-    public boolean isAllowed(ZoneBase zone, Player attacker, Entity defender, int damage);
+    public boolean isAllowed(ZoneBase zone, Player attacker, Entity defender, double damage);
 }
